fix: fix enable button test
