+++ app/src/androidTest/java/com/github/lookupgroup27/lookup/ui/googlemap/GoogleMapKtTest.kt
@@ -61,15 +61,13 @@ class GoogleMapScreenTest {
 
     // Pass the mock MutableState to LocationProvider
     locationProvider = LocationProvider(context, mockCurrentLocation)
-
-    // Set the Compose content to GoogleMapScreen
-    composeTestRule.setContent {
-      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
-    }
   }
 
   @Test
   fun mapScreenDisplaysCorrectly() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
+    }
 
     // Verify that the GoogleMapScreen is displayed
     composeTestRule.onNodeWithTag("googleMapScreen").assertIsDisplayed()
@@ -80,6 +78,10 @@ class GoogleMapScreenTest {
 
   @Test
   fun cameraPositionIsUpdatedOnLocationChange() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
+    }
+
     val fakeLocation = LatLng(37.7749, -122.4194) // Example coordinates for San Francisco
 
     // Simulate location update
@@ -101,6 +103,10 @@ class GoogleMapScreenTest {
 
   @Test
   fun buttonsAreDisplayedCorrectly() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
+    }
+
     // Verify that the "Auto Center On" button is displayed
     composeTestRule.onNodeWithText("Auto Center On").assertIsDisplayed()
 
@@ -110,6 +116,10 @@ class GoogleMapScreenTest {
 
   @Test
   fun autoCenteringButtonsFunctionality() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
+    }
+
     // Click the "Auto Center Off" button to disable auto-centering
     composeTestRule.onNodeWithText("Auto Center Off").performClick()
 
@@ -150,6 +160,9 @@ class GoogleMapScreenTest {
 
   @Test
   fun floatingActionButtonDisplaysAndNavigatesToCameraCapture() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel)
+    }
     // Mock the FirebaseAuth instance
     mockAuth = org.mockito.kotlin.mock()
     whenever(mockAuth.currentUser).thenReturn(null) // Can change this to test different scenarios
@@ -166,5 +179,19 @@ class GoogleMapScreenTest {
     } else {
       verify(navigationActions).navigateTo(Screen.AUTH)
     }
+  }
+
+  @Test
+  fun testEnableLocationButtonIsDisplayed() {
+    composeTestRule.setContent {
+      GoogleMapScreen(navigationActions, postsViewModel, profileViewModel, true)
+    }
+
+    // Verify the UI elements are displayed
+    composeTestRule.onNodeWithTag("background_image").assertExists() // Verify image is displayed
+    composeTestRule
+        .onNodeWithTag("enable_location_button")
+        .assertExists() // Verify button is displayed
+    composeTestRule.onNodeWithText("Enable Location").assertExists() // Verify button text
   }
 }
